Update CARTO map URL (#73)
AEMP made some changes to their CARTO account and so we needed to update the map url. We also updated contentful for all the individual map links for Evictors. Initially this netlify building was failing, that's becaues there was no specific python version set, and the latest one it chose to use caused some conflict with systems. I added an env variable on netlify to set python to 2.7, what was used on the last successful build (in 2022). Separately we should work on updating that to something recent before it's deprecated completely as an option